# mobile/watch-fax/swagger/apiSpec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "watch-fax-backend",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/UserCollections": {
            "post": {
                "tags": ["UserCollections"],
                "operationId": "CreateWatch",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "Brand": {
                                        "type": "string"
                                    },
                                    "Model": {
                                        "type": "string"
                                    },
                                    "ReferenceNumber": {
                                        "type": "string"
                                    },
                                    "SerialNumber": {
                                        "type": "string"
                                    },
                                    "ProductionYear": {
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "PurchaseDate": {
                                        "type": "integer",
                                        "format": "int64"
                                    },
                                    "PurchasePrice": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "HasPapers": {
                                        "type": "boolean"
                                    },
                                    "HasBox": {
                                        "type": "boolean"
                                    },
                                    "HasRecordOfAuthentication": {
                                        "type": "boolean"
                                    },
                                    "DescriptionOfCondition": {
                                        "type": "string"
                                    },
                                    "Story": {
                                        "type": "string"
                                    },
                                    "Images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                }
                            },
                            "encoding": {
                                "Brand": {
                                    "style": "form"
                                },
                                "Model": {
                                    "style": "form"
                                },
                                "ReferenceNumber": {
                                    "style": "form"
                                },
                                "SerialNumber": {
                                    "style": "form"
                                },
                                "ProductionYear": {
                                    "style": "form"
                                },
                                "PurchaseDate": {
                                    "style": "form"
                                },
                                "PurchasePrice": {
                                    "style": "form"
                                },
                                "HasPapers": {
                                    "style": "form"
                                },
                                "HasBox": {
                                    "style": "form"
                                },
                                "HasRecordOfAuthentication": {
                                    "style": "form"
                                },
                                "DescriptionOfCondition": {
                                    "style": "form"
                                },
                                "Story": {
                                    "style": "form"
                                },
                                "Images": {
                                    "style": "form"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/WatchRecord"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WatchRecord"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WatchRecord"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": ["UserCollections"],
                "operationId": "GetAllWatchRecords",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WatchRecord"
                                    }
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WatchRecord"
                                    }
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/WatchRecord"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": {}
            },
            "WatchRecord": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "nullable": true
                    },
                    "userId": {
                        "type": "string",
                        "nullable": true
                    },
                    "brand": {
                        "type": "string",
                        "nullable": true
                    },
                    "model": {
                        "type": "string",
                        "nullable": true
                    },
                    "referenceNumber": {
                        "type": "string",
                        "nullable": true
                    },
                    "serialNumber": {
                        "type": "string",
                        "nullable": true
                    },
                    "productionYear": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "purchaseDate": {
                        "type": "integer",
                        "format": "int64",
                        "nullable": true
                    },
                    "purchasePrice": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "hasPapers": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "hasBox": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "hasRecordOfAuthentication": {
                        "type": "boolean",
                        "nullable": true
                    },
                    "descriptionOfCondition": {
                        "type": "string",
                        "nullable": true
                    },
                    "story": {
                        "type": "string",
                        "nullable": true
                    },
                    "imageUrls": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    }
                },
                "additionalProperties": false
            }
        }
    }
}